requirements.txt
opensearch-py==2.4.2
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
faker==19.6.2